LoanController.php

+++ app/Http/Controllers/LoanController.php
@@ -326,8 +326,13 @@ class LoanController extends Controller
         $storeLoanAccountId = $loan->loan_account_id;
         $storedType = $loan->type;
         $storedCompanyId = $loan->loan_company_id;
+
         if ($loan->total_paid > 0) {
-            return response()->json(['errorMsg' => 'This loan can not delete. Some or full amount has been paid/received on this loan.']);
+            return response()->json(['errorMsg' => 'This loan can not delete. Some or full amount has been paid on this loan.']);
+        }
+
+        if ($loan->total_receive > 0) {
+            return response()->json(['errorMsg' => 'This loan can not delete. Some or full amount has been received on this loan.']);
         }
 
         $loan->delete();
